refactor test mocks to share across modules
