+++ tests/samples/test_target_csv.py
@@ -14,9 +14,6 @@ from freezegun import freeze_time
 from samples.sample_mapper.mapper import StreamTransform
 from samples.sample_tap_countries.countries_tap import SampleTapCountries
 from samples.sample_target_csv.csv_target import SampleTargetCSV
-from singer_sdk import typing as th
-from singer_sdk.sinks import BatchSink
-from singer_sdk.target_base import Target
 from singer_sdk.testing import (
     get_target_test_class,
     sync_end_to_end,
@@ -24,6 +21,7 @@ from singer_sdk.testing import (
     tap_to_target_sync_test,
     target_sync_test,
 )
+from tests.conftest import TargetMock 
 
 TEST_OUTPUT_DIR = Path(f".output/test_{uuid.uuid4()}/")
 SAMPLE_CONFIG = {"target_folder": f"{TEST_OUTPUT_DIR}/"}
@@ -53,55 +51,6 @@ SAMPLE_TAP_CONFIG: dict[str, t.Any] = {}
 COUNTRIES_STREAM_MAPS_CONFIG: dict[str, t.Any] = {
     "stream_maps": {"continents": {}, "__else__": None},
 }
-
-
-class BatchSinkMock(BatchSink):
-    """A mock Sink class."""
-
-    name = "batch-sink-mock"
-
-    def __init__(
-        self,
-        target: TargetMock,
-        stream_name: str,
-        schema: dict,
-        key_properties: list[str] | None,
-    ):
-        """Create the Mock batch-based sink."""
-        super().__init__(target, stream_name, schema, key_properties)
-        self.target = target
-
-    def process_record(self, record: dict, context: dict) -> None:
-        """Tracks the count of processed records."""
-        self.target.num_records_processed += 1
-        super().process_record(record, context)
-
-    def process_batch(self, context: dict) -> None:
-        """Write to mock trackers."""
-        self.target.records_written.extend(context["records"])
-        self.target.num_batches_processed += 1
-
-
-class TargetMock(Target):
-    """A mock Target class."""
-
-    name = "target-mock"
-    config_jsonschema = th.PropertiesList().to_dict()
-    default_sink_class = BatchSinkMock
-
-    def __init__(self):
-        """Create the Mock target sync."""
-        super().__init__(config={})
-        self.state_messages_written: list[dict] = []
-        self.records_written: list[dict] = []
-        self.num_records_processed: int = 0
-        self.num_batches_processed: int = 0
-
-    def _write_state_message(self, state: dict):
-        """Emit the stream's latest state."""
-        super()._write_state_message(state)
-        self.state_messages_written.append(state)
-
 
 def test_countries_to_csv(csv_config: dict):
     tap = SampleTapCountries(config=SAMPLE_TAP_CONFIG, state=None)
@@ -133,7 +82,7 @@ def test_target_batching():
     countries_record_count = 257
 
     with freeze_time(mocked_starttime):
-        target = TargetMock()
+        target = TargetMock(config={})
         target.max_parallelism = 1  # Limit unit test to 1 process
         assert target.num_records_processed == 0
         assert len(target.records_written) == 0
